backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0,<4.0
djangorestframework-simplejwt>=5.3.1,<6.0
django-cors-headers>=4.3.1,<5.0
django-filter>=24.2,<25.0
python-dotenv>=1.0.1,<2.0
Pillow>=10.3.0,<11.0
openpyxl>=3.1.5,<4.0
gunicorn>=23.0.0,<24.0
whitenoise>=6.7.0,<7.0
dj-database-url>=2.2.0,<3.0
psycopg2-binary>=2.9.9,<3.0
